+++ web/CHANGES.md
@@ -1,1 +1,4 @@
 # unreleased
+
+## Add
+- implement `Responder` trait for `serde_json::Value`.